requirements.txt
django==4.0.3
gdal==3.4.1
pip-chill==1.0.1
psycopg2==2.9.3
pytelegrambotapi==4.4.0
python-decouple==3.6
